requirements.txt
pydantic
pyyaml
PyQt6
requests
